+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - python/utils: return axes instances from velocity/controls plotting helpers
 - make `LinearFuntionTpl::evaluate()` call more efficient (using `.noalias()`)
 - `HistoryCallbackTpl` now stores stored data directly
+- Deprecate the `StageConstraintTpl` template struct, deprecate the related typedefs
 
 ### Removed